Revised handling for Model document fields
